Moved queryParameter.js file to common folder
